+++ src/renderers/post/elements/index.ts
@@ -5,3 +5,4 @@ export * from "./BloomElement.js";
 export * from "./QuantizeElement.js";
 export * from "./HbaoElement.js";
 export * from "./TaaElement.js";
+export * from "./MotionTrailElement.js";