+++ src/renderers/post/passes/index.ts
@@ -3,5 +3,5 @@ export * from "./BloomPassGPU.js";
 export * from "./PostProcessPassGL.js";
 export * from "./AOPassGL.js";
 export * from "./AOPassGPU.js";
-export * from "./TAAPassGL.js";
-export * from "./TAAPassGPU.js";
+export * from "./HistoryBlendPassGL.js";
+export * from "./HistoryBlendPassGPU.js";